+++ server/app/models/audit_log.py
@@ -49,6 +49,8 @@ class AuditLog(db.Model):
     ACTION_AUDIT_LOGS_ACCESS = 'audit_logs_access'
     ACTION_AUDIT_LOGS_EXPORT = 'audit_logs_export'
     ACTION_PASSWORD_CHANGE = 'password_change'
+    ACTION_PASSWORD_RESET_REQUEST = 'password_reset_request'
+    ACTION_PASSWORD_RESET = 'password_reset'
     
     @classmethod
     def log(cls, action, user_id=None, target_type=None, target_id=None,